refactor: 💡 file name typo
indes -> index
